requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.15.1
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
install==1.3.5
pycparser==2.21
pytz==2023.3
sqlparse==0.4.4
